added activity create/edit pages

+++ karma-api/src/main/java/com/jshimas/karmaapi/services/impl/AuthTokenServiceImpl.java
@@ -35,7 +35,7 @@ public class AuthTokenServiceImpl implements AuthTokenService {
     private final JwtEncoder jwtEncoder;
 
     @Value("${jwt.expiration.access}")
-    private int accessExpirationMs;
+    private long accessExpirationMs;
 
     @Value("${jwt.expiration.refresh}")
     private long refreshExpirationMs;
